Move business logic from datasource to use case
