+++ timetable/Data/EventsDataSource.swift
@@ -5,8 +5,9 @@
 //  Created by Hoff Silva on 03/06/21.
 //
 
+import EventKit
 
 protocol EventsDataSource {
-    func getEvents(from: Int, completion: @escaping ((Result<[Month], Error>) -> Void))
+    func getEvents(from: Int, completion: @escaping ((Result<[EKEvent], Error>) -> Void))
     func requestAccess(completion: @escaping ((Bool) -> Void))
 }
