+++ timetable/Domain/Repositories/GetEventsRepository.swift
@@ -5,7 +5,9 @@
 //  Created by Hoff Silva on 03/06/21.
 //
 
+import EventKit
+
 protocol GetEventsRepository {
-    func getEvents(from: Int, completion: @escaping ((Result<[Month], Error>) -> Void))
+    func getEvents(from: Int, completion: @escaping ((Result<[EKEvent], Error>) -> Void))
     func requestAccess(completion: @escaping ((Bool) -> Void))
 }
